requirements.txt
Django==3.2.9
djangorestframework==3.12.4
coverage==6.2
django-cors-headers==3.10.1
djangorestframework-simplejwt==5.0.0
PyYAML==6.0
uritemplate==4.1.1
coreapi==2.3.3
